Добавлена проверка полей форма и исправлено авторизация пользователя

+++ user/routers.py
@@ -1,9 +1,8 @@
 from fastapi import APIRouter, HTTPException, Depends, status, Response
 from sqlalchemy.exc import IntegrityError
-from fastapi.security import OAuth2PasswordRequestForm
 from user.my_token import create_access_token, NAME_COOKIES
 from src.api_models import User
-from user.schemas import UserSchema, AdminUserScheme
+from user.schemas import UserSchema, AdminUserScheme, AuthUserScheme
 from api_databases.connect_db import get_async_session, data_is_not_valid
 from sqlalchemy.ext.asyncio import AsyncSession
 from passlib.hash import pbkdf2_sha256
@@ -15,25 +14,28 @@ router = APIRouter(
 )
 
 
+async def field_validation(field: dict) -> list:
+    """Валидация полей формы"""
+    errors_list = list()
+    for key in field.keys():
+        if isinstance(field[key], dict):
+            for value in field[key].values():
+                errors_list.append(value)
+    return errors_list
+
+
 @router.post("/user_create")
 async def user_create(user: UserSchema = Depends(UserSchema.as_form),
                       session: AsyncSession = Depends(get_async_session)
                       ):
     """Создание пользователя"""
-    errors_list = list()
     user_data = user.dict()
+    # Проверка наличия ошибок валидации полей
+    errors_list = await field_validation(user_data)
+    if errors_list:
+        return {"errors": errors_list}
     msg = 'A user with the same name or email already exists'
     try:
-        # Проверка наличия ошибок валидации полей
-        if isinstance(user_data["username"], dict):
-            errors_list.append(user_data["username"]["message"])
-        if isinstance(user_data["password"], dict):
-            errors_list.append(user_data["password"]["message"])
-        if isinstance(user_data["email"], dict):
-            errors_list.append(user_data["email"]["message"])
-        if errors_list:
-            return {"errors": errors_list}
-
         # Проверка наличия пользователя в базе данных с таким именем
         query = select(User).filter(User.username == user_data["username"].title())
         result = await session.execute(query)
@@ -72,25 +74,35 @@ async def user_create(user: UserSchema = Depends(UserSchema.as_form),
 @router.post("/login")
 async def login(
         response: Response,
-        request: OAuth2PasswordRequestForm = Depends(),
+        user_form: AuthUserScheme = Depends(AuthUserScheme.as_form),
         session: AsyncSession = Depends(get_async_session)
 ):
     """Авторизация пользователя"""
-    query = select(User).filter(User.username == request.username)
-    user = await session.scalar(query)
+    user_data = user_form.dict()
+    # Проверка наличия ошибок валидации полей
+    errors_list = await field_validation(user_data)
+    if errors_list:
+        return {"errors": errors_list}
+
     # Аутентификация пользователя
+    query = select(User).filter(User.username == user_data["username"].title())
+    user = await session.scalar(query)
     if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User: {request.username} not found'")
-    elif not pbkdf2_sha256.verify(request.password, user.password):
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Password: '{request.password}' invalid password"
-        )
+        msg = f"User: {user_data['username']} not found'"
+        errors_list.append(msg)
+        return {"errors": errors_list}
+    elif not pbkdf2_sha256.verify(user_data["password"], user.password):
+        msg = f"Password: '{user_data['password']}' invalid password"
+        errors_list.append(msg)
+        return {"errors": errors_list}
     # Создание токена
     jwt_token = create_access_token(data={"sub": user.username})
     # Сохранение токена в cookie
     response.set_cookie(key=NAME_COOKIES, value=f"Bearer {jwt_token}", httponly=True)
-    return {"message": f"{user.username} -> login"}
+    return {
+        "token": jwt_token,
+        "username": user.username,
+    }
 
 
 @router.patch("/update_user_group/{user_id}", status_code=status.HTTP_202_ACCEPTED)
@@ -119,19 +131,20 @@ async def delete_user(user_id: int, current_user: dict = Depends(get_current_use
                       session: AsyncSession = Depends(get_async_session)
                       ):
     """Удаление пользователя"""
+    if current_user is None:
+        return data_is_not_valid
     if current_user["group"] == "ADMIN" or current_user["user_id"] == user_id:
         try:
             query = select(User).filter(User.id == user_id)
             user_delete = await session.execute(query)
             await session.delete(user_delete.scalar())
             await session.commit()
-            response = {
+            return {
                 "status": status.HTTP_204_NO_CONTENT,
                 "detail": "User Deleted!"
             }
-            return response
+
         except Exception:
-            await session.rollback()
             raise data_is_not_valid
     else:
         raise HTTPException(
@@ -141,7 +154,8 @@ async def delete_user(user_id: int, current_user: dict = Depends(get_current_use
 
 
 @router.put("/update_user/{user_id}")
-async def update_user(user_id: int, user: UserSchema, current_user: dict = Depends(get_current_user),
+async def update_user(user_id: int, user: UserSchema = Depends(UserSchema.as_form),
+                      current_user: dict = Depends(get_current_user),
                       session: AsyncSession = Depends(get_async_session)
                       ):
     """Обновление пользователя"""
@@ -188,9 +202,8 @@ async def update_user(user_id: int, user: UserSchema, current_user: dict = Depen
     )
 
 
-@router.post("/logout")
+@router.get("/logout")
 async def user_logout(response: Response, current_user: dict = Depends(get_current_user)):
     """Выход пользователя из приложения"""
-    username = current_user["username"]
     response.delete_cookie(NAME_COOKIES)
-    return {"message": f"Пользователь {username} -> вышел из сети"}
+    return {"message": f'User: {current_user["username"]} logged out'}
